+++ examples/ca-opportunity-example/src/ca_common_grants/services/opportunity.py
@@ -17,7 +17,7 @@ from common_grants_sdk.schemas.models import (
 )
 from fastapi import status
 
-from ca_common_grants.transform.transformer import CATransformer
+from ca_common_grants.utils.opp_map_transform import OpportunityMapTransformer
 
 DATA_FILE = Path(__file__).parent.parent / "data" / "ca_grants_sample.json"
 
@@ -39,7 +39,11 @@ class OpportunityService:
 
     def _get_opportunity_list(self) -> list[OpportunityBase]:
         """Transform CA opportunity data to CGP format."""
-        transformed: list[dict[str, Any]] = CATransformer.from_file(DATA_FILE)
+        transformed: list[dict[str, Any]] = (
+            OpportunityMapTransformer.transform_opportunities_file(
+                DATA_FILE,
+            )
+        )
 
         """Build result set."""
         return [OpportunityBase.model_validate(opp_data) for opp_data in transformed]
