--- examples/ca-opportunity-example/src/ca_common_grants/transform/__init__.py
@@ -1,1 +1,0 @@
-"""Transform package for converting CA Grants Portal data to CommonGrants Protocol format."""
