+++ examples/ca-opportunity-example/src/ca_common_grants/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for working with CA Grants Portal data and CommonGrants Protocol."""
